添加template
